Support inheritance from Kind
Kind expects a &Kind::set_parameterizer call at BEGIN-time within the
body of its child's package. This prepares any given metaobject for
parameterization via ^kind on the child's HOW. This would typically be
the child, but may be a package of sorts should any changes to the
parameterizer need to be made. It can accept an optional parameterizer
to replace &Kind::parameterizer if need be.

Kind.^kind is equivalent to ^parameterize on Kind itself, but generally
should not be overridden. Added functionality makes for added
metamethods or an override of ^parameterize instead.